rename evm to vm
