+++ src/ethereum/vm/error.py
