+++ src/ethereum/vm/ops.py
