+++ src/ethereum/vm/stack.py
