+++ whitelist.txt
@@ -53,6 +53,7 @@ fromhex
 preimage
 substring
 klass
+vm
 
 sha3
